refactor(extractor): move hurt sound into entity extraction
