+++ src/main/kotlin/de/snowii/extractor/Extractor.kt
@@ -39,7 +39,6 @@ class Extractor : ModInitializer {
             PotionBrewing(),
             Potion(),
             Sounds(),
-            EntityHurtSound(),
             Recipes(),
             Biome(),
             BiomeMixerTest(),
