+++ src/main/kotlin/de/snowii/extractor/extractors/Entities.kt
@@ -6,7 +6,9 @@ import com.google.gson.JsonObject
 import com.mojang.serialization.JsonOps
 import de.snowii.extractor.Extractor
 import net.minecraft.core.registries.BuiltInRegistries
+import net.minecraft.resources.Identifier
 import net.minecraft.server.MinecraftServer
+import net.minecraft.world.damagesource.DamageSource
 import net.minecraft.world.entity.EntitySpawnReason
 import net.minecraft.world.entity.LivingEntity
 import net.minecraft.world.entity.Mob
@@ -23,15 +25,25 @@ class Entities : Extractor.Extractor {
         val entitiesJson = JsonObject()
         val registryAccess = server.registries().compositeAccess()
         val ops = registryAccess.createSerializationContext(JsonOps.INSTANCE)
+        val damageSource = server.overworld().damageSources().generic()
 
         for (entityType in BuiltInRegistries.ENTITY_TYPE) {
             val entityJson = JsonObject()
+            val entityName = BuiltInRegistries.ENTITY_TYPE.getKey(entityType).path
             entityJson.addProperty("id", BuiltInRegistries.ENTITY_TYPE.getId(entityType))
 
             val entity = entityType.create(server.overworld(), EntitySpawnReason.NATURAL)
             if (entity != null) {
                 if (entity is LivingEntity) {
                     entityJson.addProperty("max_health", entity.maxHealth)
+
+                    if (entityName in TARGET_HURT_SOUND_ENTITIES) {
+                        val hurtSound = getHurtSound(entity, damageSource)
+                        val hurtSoundId = hurtSound?.let(BuiltInRegistries.SOUND_EVENT::getKey)?.path
+                        if (hurtSoundId != null) {
+                            entityJson.addProperty("hurt_sound", hurtSoundId)
+                        }
+                    }
                 }
                 entityJson.addProperty("attackable", entity.isAttackable)
                 entityJson.addProperty("mob", entity is Mob)
@@ -75,11 +87,33 @@ class Entities : Extractor.Extractor {
             entityJson.add("spawn_restriction", spawnRestriction)
 
             entitiesJson.add(
-                BuiltInRegistries.ENTITY_TYPE.getKey(entityType).path,
+                entityName,
                 entityJson
             )
         }
 
         return entitiesJson
+    }
+
+    private fun getHurtSound(entity: LivingEntity, damageSource: DamageSource) =
+        getHurtSoundMethod.invoke(entity, damageSource) as? net.minecraft.sounds.SoundEvent
+
+    companion object {
+        private val TARGET_HURT_SOUND_ENTITIES = setOf(
+            "bogged",
+            "drowned",
+            "enderman",
+            "husk",
+            "parched",
+            "skeleton",
+            "stray",
+            "wither_skeleton",
+            "zombie",
+            "zombie_villager",
+        )
+
+        private val getHurtSoundMethod = LivingEntity::class.java
+            .getDeclaredMethod("getHurtSound", DamageSource::class.java)
+            .apply { isAccessible = true }
     }
 }
